Backend/flight_radar/_docker_context/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0

# Async HTTP
httpx==0.27.0

# Redis
redis[hiredis]==5.0.4

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Settings
pydantic==2.7.1
pydantic-settings==2.3.0

# Metrics
prometheus-client==0.20.0

# Retry / Resilience
tenacity==8.3.0

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0

# Linting / Typing
ruff==0.4.5
mypy==1.10.0
types-redis==4.6.0.20240417
